Combo Jump setted up and started with AirCombo
Inmediately after ToAir hit, the player can press the jump button to perform a special jump "towards the enemy". While the player is in this special jump a timer is checking if has passed enough time to change to InAirState. If the timer hasn't run out, the player can press the attack button when is very close to the enemy to perform an air attack starting the air combo.