PushAttack and ComboDash added
If the combo chain selected for one of the attack inputs is 2, the fourth attack performed with this input is a kick that pushes the enemy back way more than the normal attack. Then the player can do a special dash to quickly get close to the enemy and continue the combo.
The mass of the enemy was doubled and so the forces applied when is hit.
Unwanted behaviours:
- After the player do the ToAir (fromthe kick combo) attack, if the dash button is pressed, it does the combo dash anyway. The same behavior applies in the opposite way: if push attack is performed and the player hits the jump button, it does the ComboJump.