Combo Issues Fixed
The player combo is reseted if the input pressed during the current combo is from another combo type.
The player only changes to ToAirAttack if the combo type of the input pressed is equal to 2.
The player only changes to PushAttack if the combo type of the input pressed is equal to 1.
The player can Combo Dash if the last attack's combo type was 1. Likewise, the player can ComboJump if the last attack's input type was 2.
The player can AirAttack only if the pressed input's combo type is equal to 2.